requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alpaca-py==0.38.0
alpaca-trade-api==3.2.0
altgraph @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.17.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.2.0
future @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
ibapi==9.81.1.post1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
lxml==5.3.1
macholib @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.0.3
multidict==6.1.0
multitasking==0.0.11
networkx==3.2.1
numpy==2.0.2
openai==1.61.1
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
polygon==1.2.6
polygon-api-client==1.14.3
propcache==0.2.1
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.1
requests==2.32.3
scipy==1.13.1
six @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
soupsieve==2.6
sseclient-py==1.8.0
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
virtualenv==20.29.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
yarl==1.18.3
yfinance==0.2.52
